Refactor nearest_buildable action
Removed deprecated _nearest_buildable.py file and created a new nearest_buildable.py file to update the NearestBuildable class with new features and enhancements. Updated method signature and functionality to align with current requirements and improve performance and maintainability. Added new imports and modified existing code for better code organization and readability. Updated tests in test_oil.py to reflect changes in the NearestBuildable class usage and ensure proper functionality testing. Branch: feature/oil
